Draft Contributor License Agreement
Just in case LGPL doesn't fit the need in the future- I've heard it is difficult to re-license if something like this isn't put into place up front. Patterned after the GitHub CLA (https://cla.github.com/agreement) in the hope that it will be clear and preserve everyone's rights.